Remove `setup_structlog` function from module (move to tests section)
