+++ tests/conftest.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import logging
 import os
 import sys
 from collections import defaultdict
@@ -7,8 +8,7 @@ from pathlib import Path
 from typing import TYPE_CHECKING
 
 import pytest
-
-from pydiverse.pipedag.core.config import setup_structlog
+import structlog
 
 if TYPE_CHECKING:
     pass
@@ -29,6 +29,32 @@ def setup_environ():
     os.environ["MSSQL_PASSWORD"] = "PidyQuant27"
 
     os.environ["PYDIVERSE_PIPEDAG_PYTEST"] = "1"
+
+
+def setup_structlog(
+    _log_level=logging.INFO,
+    _log_stream=sys.stderr,
+    timestamp_format="%Y-%m-%d %H:%M:%S.%f",
+):
+    logging.basicConfig(
+        stream=_log_stream,
+        format="%(asctime)s [%(levelname)s] %(message)s",
+        level=_log_level,
+    )
+    structlog.configure(
+        processors=[
+            structlog.contextvars.merge_contextvars,
+            structlog.processors.add_log_level,
+            structlog.processors.StackInfoRenderer(),
+            structlog.dev.set_exc_info,
+            structlog.processors.TimeStamper(fmt=timestamp_format),
+            structlog.dev.ConsoleRenderer(),
+        ],
+        wrapper_class=structlog.make_filtering_bound_logger(_log_level),
+        context_class=dict,
+        logger_factory=structlog.PrintLoggerFactory(_log_stream),
+        cache_logger_on_first_use=True,
+    )
 
 
 setup_environ()
